Implemented Application DbContext and created database using code-first approach.

+++ Job-portal-backend-net8/Core/Entities/Job.cs
@@ -12,7 +12,7 @@ namespace Job_portal_backend_net8.Core.Entities
 
         //relationship with Company model
         [ForeignKey(nameof(CompanyId))]
-        public int CompanyId { get; set; } 
+        public long CompanyId { get; set; } 
         public Company Company { get; set; }
 
         //relationship with Candidate model
